refactor(tauri.js) rename api-src/models to api-src/types
