+++ cli/tauri.js/api-src/types/event.ts
@@ -1,0 +1,6 @@
+export interface Event {
+  type: string
+  payload: unknown
+}
+
+export type EventCallback = (event: Event) => void
